Updated Fibonacci class

+++ Logic/Fibonacci/Fibonacci.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Numerics;
 
 namespace Logic
 {
@@ -9,14 +10,14 @@ namespace Logic
         /// Generates sequence of Fibonacci numbers.
         /// </summary>
         /// <param name="quantity">Quantity of numbers to generate.</param>
-        /// <returns>Enumerator to iterate.</returns>
-        public static IEnumerable<int> Generate(int quantity)
+        /// <returns>Fibonacci sequence.</returns>
+        public static IEnumerable<BigInteger> Generate(int quantity)
         {
             if (quantity <= 0) throw new ArgumentException($"{nameof(quantity)} is unsuitable.");
 
-            int prevNumber = -1;
-            int nextNumber = 1;
-            int temp;
+            BigInteger prevNumber = -1;
+            BigInteger nextNumber = 1;
+            BigInteger temp;
             for(int i = 0; i < quantity; i++)
             {
                 temp = nextNumber;
